+++ apps/www/src/modules/home/components/key-facts-animated.tsx
@@ -1,10 +1,13 @@
 "use client";
 
 import { Heading } from "@/modules/common/components/heading";
+import { Components } from "@/modules/common/types/components";
 import { motion } from "framer-motion";
 import { useMediaQuery } from "react-responsive";
 
-const KeyFactsAnimated = () => {
+export type KeyFactsProps = Components.KeyFacts;
+
+const KeyFactsAnimated: React.FC<KeyFactsProps> = ({ featured, title }) => {
   const isLarge = useMediaQuery({ minWidth: 1024 });
 
   const variants = {
@@ -25,42 +28,29 @@ const KeyFactsAnimated = () => {
     },
   };
 
-  const data = [
-    {
-      number: "8.2k",
-      text: "Trusted Users",
-    },
-    {
-      number: "120+",
-      text: "Projects Completed",
-    },
-    {
-      number: "97%",
-      text: "Clients Satisfaction",
-    },
-  ];
-
   return (
     <>
-      <motion.section
-        className="w-full lg:max-w-[530px]"
-        variants={variants}
-        initial="left"
-        whileInView="show"
-        viewport={{ once: true, margin: "-50px" }}
-        transition={{
-          duration: 0.5,
-          ease: "easeOut",
-        }}
-      >
-        <Heading
-          size="md"
-          variant="primary"
-          className="md:text-[48px] md:leading-[48px]"
+      {!!title && (
+        <motion.section
+          className="w-full lg:max-w-[530px]"
+          variants={variants}
+          initial="left"
+          whileInView="show"
+          viewport={{ once: true, margin: "-50px" }}
+          transition={{
+            duration: 0.5,
+            ease: "easeOut",
+          }}
         >
-          GET AQUANTED WITH OUR KEY FACTS
-        </Heading>
-      </motion.section>
+          <Heading
+            size="md"
+            variant="primary"
+            className="md:text-[48px] md:leading-[48px]"
+          >
+            {title}
+          </Heading>
+        </motion.section>
+      )}
       <div>
         <motion.ul
           className="md:flex text-left flex-col items-start  md:flex-row md:justify-evenly relative gap-5 flex-wrap"
@@ -69,13 +59,13 @@ const KeyFactsAnimated = () => {
           variants={variants}
           viewport={{ once: true, margin: "0px" }}
         >
-          {data.map((_item, index) => (
+          {featured.map((_item, index) => (
             <motion.li className="text-center" variants={variants} key={index}>
               <h3 className="font-semibold text-5xl leading-[48px] mb-2">
-                {_item.number.slice(0, -1)}
-                <span className="text-primary">{_item.number.slice(-1)}</span>
+                {_item.text}
+                <span className="text-primary">{_item.featured}</span>
               </h3>
-              <p className="text-base leading-6 text-white-56">{_item.text}</p>
+              <p className="text-base leading-6 text-white-56">{_item.muted}</p>
             </motion.li>
           ))}
         </motion.ul>