+++ apps/www/src/modules/home/widgets/about-us.tsx
@@ -1,12 +1,13 @@
 import { Container } from "@components/container";
 import { Layout } from "@components/layout";
 import AboutUsAnimated, {
-  type AboutUsProps,
+  AboutUsAnimatedProps,
 } from "../components/about-us-animated";
+
+type AboutUsProps = AboutUsAnimatedProps;
 
 const AboutUs: React.FC<AboutUsProps> = (props) => {
   const { with_container: container } = props;
-
   return (
     <Container enabled={container}>
       <Layout>