+++ apps/api/CHANGELOG.md
@@ -1,5 +1,29 @@
 
 
+## [0.4.0](https://github.com/zitdevs/deft-dodo/compare/@deft-dodo/api-v0.3.0...@deft-dodo/api-v0.4.0) (2024-03-09)
+
+
+### Features
+
+* **api:** add new export file and update package dependencies ([342758b](https://github.com/zitdevs/deft-dodo/commit/342758bd0e169b287d471b6fe7e13d287463a2d6))
+
+## [0.3.0](https://github.com/zitdevs/deft-dodo/compare/@deft-dodo/api-v0.2.2...@deft-dodo/api-v0.3.0) (2024-03-09)
+
+
+### Features
+
+* **api:** add Key Facts and Featured Item components ([ad09a7d](https://github.com/zitdevs/deft-dodo/commit/ad09a7dd75bec3c0dfb896e19a08a807f6f88313))
+* **api:** DDT-33 add margin attributes to separator component ([39c7cf9](https://github.com/zitdevs/deft-dodo/commit/39c7cf9b576eec8dbb295ca5bdd0df3d1a1da511))
+* **api:** DDT-33 introduce common spacer component and update home heading structure ([dd9694a](https://github.com/zitdevs/deft-dodo/commit/dd9694a9178873006367e44a5dad4dd41378ab86))
+* **api:** DDT-33 update entities ([63bde6f](https://github.com/zitdevs/deft-dodo/commit/63bde6f14c712d561165de20d431bca47931dacd))
+* **api:** enhance service and home services schema with new fields ([16dfaba](https://github.com/zitdevs/deft-dodo/commit/16dfaba5c10a6c2c6e42cb1789228c3069caf407))
+* **api:** update type definitions for home services and integrate CMS analyzer plugin ([76a6c61](https://github.com/zitdevs/deft-dodo/commit/76a6c611c4033c30f47c6b7ef9f920d461d88302))
+
+
+### Bug Fixes
+
+* **api:** update and remove export files ([944eab2](https://github.com/zitdevs/deft-dodo/commit/944eab294adb9d83a8662c12f7b30c8a7e9e1647))
+
 ## [0.2.2](https://github.com/zitdevs/deft-dodo/compare/@deft-dodo/api-v0.2.1...@deft-dodo/api-v0.2.2) (2024-02-21)
 
 ## [0.2.1](https://github.com/zitdevs/deft-dodo/compare/@deft-dodo/api-v0.2.0...@deft-dodo/api-v0.2.1) (2024-02-17)